인터페이스 정리 및 인터페이스를 기반으로 schema 및 dto 정리

+++ src/posts/posts.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { NotFoundPostException } from '../exceptions/NotFoundPostException';
 import { InjectModel } from '@nestjs/mongoose';
-import { BPost, PostDocument } from './schema/post.schema';
+import { PostModel, PostDocument } from './schema/post.schema';
 import { Model } from 'mongoose';
 import { QueryPostDto } from './dto/query-post.dto';
 import { CreatePostDto } from './dto/create-post.dto';
@@ -11,29 +11,30 @@ import { UploadsService } from '../uploads/uploads.service';
 @Injectable()
 export class PostsService {
   constructor(
-    @InjectModel(BPost.name) private postModel: Model<PostDocument>,
+    @InjectModel(PostModel.name) private postModel: Model<PostDocument>,
     private readonly uploadsService: UploadsService
-  ) {}
+  ) {
+  }
 
   async findAll(query: QueryPostDto): Promise<PostDocument[]> {
     const { limit, orderByCreatedAt } = query;
     delete query.limit;
     delete query.orderByCreatedAt;
     return this.postModel
-      .find(query)
-      .populate({ path: 'user', select: 'name' })
-      .populate({
-        path: 'comments',
-        select: 'user content',
-        populate: { path: 'user', select: 'name' }
-      })
-      .populate({
-        path: 'attachments',
-        select: 'key'
-      })
-      .limit(limit)
-      .sort(orderByCreatedAt)
-      .exec();
+    .find(query)
+    .populate({ path: 'creator', select: 'name' })
+    .populate({
+      path: 'comments',
+      select: 'creator content',
+      populate: { path: 'creator', select: 'name' }
+    })
+    .populate({
+      path: 'attachments',
+      select: 'key'
+    })
+    .limit(limit)
+    .sort(orderByCreatedAt)
+    .exec();
   }
 
   async findById(id: string): Promise<PostDocument> {
@@ -47,7 +48,7 @@ export class PostsService {
   async create(dto: CreatePostDto): Promise<PostDocument> {
     const postDocument: PostDocument = await this.postModel.create(dto);
     await this.uploadsService.insertRef(dto.attachments || [], postDocument._id, 'Post');
-    return this.postModel.create(dto);
+    return postDocument;
   }
 
   async update(postId: string, dto: UpdatePostDto): Promise<PostDocument> {
@@ -56,19 +57,23 @@ export class PostsService {
       throw new NotFoundPostException(postId);
     }
 
-    await this.uploadsService.updateRef(dto.attachments, postId, 'Post');
-    await post.updateOne({ $set: dto }).exec();
+    await this.uploadsService.updateRef(dto.attachments || [], postId, 'Post');
+    await post.updateOne({ $set: dto });
+    await post.updateOne({ updatedAt: new Date().toString() });
     return this.findById(postId);
   }
 
-  async deleteById(id: string): Promise<PostDocument> {
-    const post: PostDocument = await this.findById(id);
+  async deleteById(postId: string): Promise<PostDocument> {
+    const post: PostDocument = await this.postModel.findById(postId);
+    if (!post) {
+      throw new NotFoundPostException(postId);
+    }
     await post.deleteOne();
     return post;
   }
 
   async deleteAll(): Promise<PostDocument[]> {
-    const posts: PostDocument[] = await this.findAll({});
+    const posts: PostDocument[] = await this.postModel.find({});
     await this.postModel.deleteMany({});
     return posts;
   }
